requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.7
Django==4.1.3
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.6